# Various performance improvements around hitobjects