1. Using list with post-sort should help with construction of any number of nested hitobjects:

|             Method | Items |         Mean |      Error |     StdDev |
| ------------------- |------ |-------------:|-----------:|-----------:|
| SortedListAddition |     5 |     88.88 us |  0.3415 us |  0.3194 us |
|           PostSort |     5 |     35.05 us |  0.1159 us |  0.1028 us |
| SortedListAddition |    10 |    196.72 us |  0.6414 us |  0.5686 us |
|           PostSort |    10 |     49.64 us |  0.1102 us |  0.1030 us |
| SortedListAddition |   100 |  2,332.01 us |  8.7999 us |  7.8009 us |
|           PostSort |   100 |    576.11 us |  3.7329 us |  3.3091 us |
| SortedListAddition |  1000 | 27,713.55 us | 64.9245 us | 57.5539 us |
|           PostSort |  1000 |  8,373.19 us | 33.7896 us | 31.6068 us |

```
public class SortBenchmarks
{
    private List<int> list = new List<int>();

    [Params(5, 10, 100, 1000)]
    public int Items { get; set; }

    [Benchmark]
    public void SortedListAddition()
    {
        for (int it = 0; it < 1000; it++)
        {
            list.Clear();
            for (int i = 0; i < Items; i++)
                addInternal(i);
        }
    }

    [Benchmark]
    public void PostSort()
    {
        for (int it = 0; it < 1000; it++)
        {
            list.Clear();
            for (int i = 0; i < Items; i++)
                list.Add(i);
            list.Sort();
        }
    }

    // Note: Replicates what SortedList<T> does
    private void addInternal(int value)
    {
        int index = list.BinarySearch(value);
        if (index < 0)
            index = ~index;
        list.Insert(index, value);
    }
}
```

2. Size of `List<T>` is smaller than `Lazy<List<T>>` without constructing the value of the lazy:

|               Method |     Mean |    Error |   StdDev |   Median |    Gen 0 | Allocated |
| --------------------- |---------:|---------:|---------:|---------:|---------:|----------:|
|     ListConstruction | 113.1 us | 7.306 us | 21.54 us | 110.4 us | 126.9531 | 390.63 KB |
| LazyListConstruction | 175.7 us | 7.341 us | 20.09 us | 167.7 us | 228.7598 | 703.13 KB |

```
[MemoryDiagnoser]
public class ConstructionBenchmarks
{
    [Benchmark]
    public void ListConstruction()
    {
        for (int i = 0; i < 10000; i++)
        {
            new List<int>();
        }
    }

    [Benchmark]
    public void LazyListConstruction()
    {
        for (int i = 0; i < 10000; i++)
        {
            new Lazy<List<int>>();
        }
    }
}
```