@jpsouzasilva 2018-10-10
Is it still better even when the elements are not already inserted in a sorted order?

Could the tests become more realistic by generating a a couple of randomly ordered, non-repeating sets before running the batch? Or is it the case that hitobjects often get appended into the list in a sorted order, like the test code?