@smoogipoo 2018-10-10
In most cases nested hitobjects are appended in sorted order.